refactor: Move DbSequence, SqlHints, and LockBehaviorBase to the root namespace
Third wave of #244, refining the boundary rule with the owner: values,
items, and handles move out of Internal; clause syntax stays. A clause
fragment is specified through the builder chain at the call site, and
the natural unit of reuse is the completed expression — already
nameable via the moved roots — so OrderByClause, PartitionByClause,
OfClause, SeparatorClause and friends remain Internal deliberately.

The three moved types are values/handles with cross-statement
identity, each with a naming flow no root type can express:

- DbSequence — a schema-object handle centralized like a table class
  (`static readonly DbSequence UserSeq = Sequence(...)`); the field
  declaration requires the type name
- SqlHints — a hint block shared across statements the same way
- LockBehaviorBase — Nowait/SkipLocked/Wait return three different
  concrete types, so runtime lock selection held in a variable or
  parameter cannot be written without the base name

The rule wording is recorded in CLAUDE.md and ADR 0005; the CHANGELOG
entry now carries all ten moved types. PublicSurfaceNamingTests gains
the three flows (shared handle field, shared hint field, runtime lock
selection), still with no `using SqlArtisan.Internal;`.

Refs #244
