+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Opt-in Roslyn analyzer, shipped inside the `SqlArtisan` package with no extra reference. Set a target dialect — `sqlartisan_target_dbms` in `.editorconfig`, or `<SqlArtisanTargetDbms>` as an MSBuild property — and `SQLA0001` warns when a construct isn't supported there, per a curated, primary-source-verified dialect matrix (silent, and no false positive, for anything the matrix hasn't confirmed one way or the other). `SQLA0002` flags an unrecognized `.editorconfig` value. Every warning names the `sqlartisan_construct_<member>[_arity<N>]` override key that would silence it (`supported`) or force it (`unsupported`) — the arity-level key exists because a handful of overloaded members (e.g. `StringAgg`'s 3-argument inline-`ORDER BY` form) have narrower dialect support than their siblings. Severity is controlled the standard way (`dotnet_diagnostic.SQLA0001.severity`), so a mismatch can be promoted to a build error. See `docs/analyzer.md`. (#93)
 ### Changed
 - **Breaking:** a query with a row-limiting clause (`Limit()` / `Offset()` / `OffsetRows()` / `FetchNext()` / `FetchFirst()`) can now be used as a subquery: a `LATERAL` / `APPLY` operand (per-group top-N), an aliased scalar select item (`.As("alias")`), a CTE body, or an `IN` / `NOT IN` / `EXISTS` / `ALL` / `ANY` / `SOME` operand. Source-compatible for callers but binary-breaking — the terminal `Offset()` / `FetchFirst()` / `FetchNext()` steps now return `ISelectBuilderPaginated` instead of `ISqlBuilder` — so rebuild code compiled against an earlier version. Recompiling also changes `In(...)` / `NotIn(...)` with a row-limited subquery, by design: previously they silently emitted a scalar-subquery comparison, `IN ((SELECT … LIMIT :0))` — wrong semantics that fail at runtime past one row — and now emit the list form `IN (SELECT … LIMIT :0)`. MySQL itself rejects `LIMIT` directly inside an `IN` / `ALL` / `ANY` / `SOME` subquery — route the limited query through a CTE there (see the pagination reference). (#240)
-- **Breaking:** the types ordinary application code must *name* — a helper method's return type, a `WHERE` / sort-key accumulator variable, a `List<>` of tables or CTEs — moved from `SqlArtisan.Internal` to the root `SqlArtisan` namespace: `SqlCondition`, `SqlExpression`, `ISubquery`, `TableReference`, `SortOrder`, `ExpressionAlias`, and `CommonTableExpression`. Previously the standard dynamic-query idioms (a nullable condition accumulator, a reusable-subquery helper, a dynamic `FROM` / `WITH` list, a UI-driven sort key) forced an implementation-detail `using SqlArtisan.Internal;` just to name them. Every concrete node (`AndCondition`, `LikeCondition`, `CountFunction`, …) stays in `Internal` and is held only through these roots; code that named the moved types fully qualified needs a one-line fix, and a now-unnecessary `using SqlArtisan.Internal;` is harmless. `CommonTableExpression`'s constructor is now `internal` — obtain it via `Cte` / `CteBase`'s `.As(...)`, the path that also yields the column-reference handle. (#244)
+- **Breaking:** the types ordinary application code must *name* — a helper method's return type, an accumulator variable, a `List<>` element, a shared `static readonly` field — moved from `SqlArtisan.Internal` to the root `SqlArtisan` namespace: `SqlCondition`, `SqlExpression`, `ISubquery`, `TableReference`, `SortOrder`, `ExpressionAlias`, `CommonTableExpression`, `DbSequence`, `SqlHints`, and `LockBehaviorBase`. Previously the standard dynamic-query idioms (a nullable condition accumulator, a reusable-subquery helper, a dynamic `FROM` / `WITH` list, a UI-driven sort key, a shared sequence handle or hint block, a runtime-selected lock behavior) forced an implementation-detail `using SqlArtisan.Internal;` just to name them. The line is values, items, and handles versus clause syntax: clause fragments (`OrderBy(...)` for `OVER` / string aggregation, `Of(...)`, `Separator(...)`) stay in `Internal` — they are specified through the builder chain, and the reusable unit is the completed expression, which the moved roots already name. Every concrete node (`AndCondition`, `LikeCondition`, `CountFunction`, …) stays in `Internal` and is held only through these roots; code that named the moved types fully qualified needs a one-line fix, and a now-unnecessary `using SqlArtisan.Internal;` is harmless. `CommonTableExpression`'s constructor is now `internal` — obtain it via `Cte` / `CteBase`'s `.As(...)`, the path that also yields the column-reference handle. (#244)
 ### Fixed
 - Corrected the `ALL` / `ANY` / `SOME` dialect claim: 0.5.0-beta.2's "supported on all five dialects" was wrong for SQLite, whose grammar has no quantified comparisons at all (`near "ALL": syntax error`) — caught by the new dialect sweep, which executes every analyzer-matrix entry against the live engines in both directions. The expression reference and the XML docs now name the four supporting dialects; emitted SQL is unchanged. (#93)
 - Synced the dialect notes in the XML docs and the reference pages with the live-verified matrix — a mechanical audit found 19 remarks still carrying pre-verification claims: the `REGEXP_*` and `TO_*` families said “Oracle” though MySQL and/or PostgreSQL support them, `INSTR`'s 2-argument form is MySQL/Oracle/SQLite, `LAST_DAY` is MySQL too, SQLite accepts `CEILING` and `STRING_AGG` (3.44+), `JSON_VALUE` is MySQL 8.0.21+, the `LATERAL` joins are Oracle 12c+, `OFFSET … FETCH` includes PostgreSQL, `WHEN MATCHED THEN DELETE` is PostgreSQL 15+, and SQL Server accepts the legacy `WITH ROLLUP` form. Emitted SQL is unchanged. (#93)
