+++ CLAUDE.md
@@ -95,13 +95,17 @@ there, not here — a pointer line in this list is enough.
   and their base `TableReference`), `DbColumn` under
   `src/SqlArtisan/SqlPart/Expression/`, and the types users must **name** in a
   declaration position (a helper's return type, an accumulator variable, a
-  `List<>` element) — `SqlExpression`, `SqlCondition`, `ISubquery`,
-  `SortOrder`, `ExpressionAlias`, `CommonTableExpression` — all in the root
-  namespace under `src/SqlArtisan/SqlPart/` / `src/SqlArtisan/SqlBuilder/`.
-  That naming test is the boundary rule (#244): a type moves out of
-  `Internal/` when a mainstream flow must write its name; every concrete node
-  stays `Internal/` and is held only through these roots. Everything else
-  under `Internal/` is implementation detail.
+  `List<>` element, a shared `static readonly` field) — `SqlExpression`,
+  `SqlCondition`, `ISubquery`, `SortOrder`, `ExpressionAlias`,
+  `CommonTableExpression`, `DbSequence`, `SqlHints`, `LockBehaviorBase` — all
+  in the root namespace under `src/SqlArtisan/SqlPart/` /
+  `src/SqlArtisan/SqlBuilder/`. The boundary rule (#244): **values, items,
+  and handles move out of `Internal/` when a mainstream flow must write
+  their name; clause syntax stays** — clause fragments are specified through
+  the builder chain, and the reusable unit is the completed expression,
+  already nameable via the roots. Every concrete node stays `Internal/` and
+  is held only through these roots. Everything else under `Internal/` is
+  implementation detail.
 - Name public members after their SQL token — **underscores are the only word
   boundaries** (`ADD_MONTHS`→`AddMonths`, `DATEADD`→`Dateadd`, never invented
   internal capitals). The full rule, the glyph/helper naming categories, and
