--- src/SqlArtisan/Internal/SqlPart/Clause/ForUpdate/LockBehaviorBase.cs
@@ -1,5 +1,0 @@
-namespace SqlArtisan.Internal;
-
-public abstract class LockBehaviorBase : SqlPart
-{
-}
